Fix warnings (#16)
* Reorder ROOTMultiPlot()

* Reorder AxisCollection()

* Remove superfluous leading consts in SystematicManager methods

* Reorder CutLog(cutNames_)

* Reorder BinnedNLLH()

* Fix BinnedNLLH (un)signed warnings; modify for-loops to be range-based

* Update ChiSquare::RegisterFitComponents() to use range-based for loops

* Fix Function's GetNDims() type; propagate to Gaussian.

* Fix constness in SystematicManager.cpp

* Fix GaussianFitter signed/unsigned warnings

* Reorder FitResult()

* Reorder MetropolisHastings()

* Remove irrelevant params in JumpProbRatio()

* Correct TestStatistic::GetParameterCount() to size_t; propagate

* Remove pointless JumpProbRatio() method

* Remove unused nParams in GridSearch::Optimise()

* Reorder Minuit()

* Remove unused variables in Convolution::Construct()

* Make fNBins a size_t in BinAxis

* Correct Histogram::AddPadding() signed/unsigned warning

* Fix signed/unsigned warnings in SystematicManager.cpp

* Change type of fNRows/fNCols in Sparse & DenseMatrix to size_t

* Fix signed/unsigned warning in DenseMatrix::SetSymmetricMatrix()

* Remove unused in IO::SaveDataSet()

* Deal with one final signed/unsigned in DenseMatrix.cpp

* Because ROOT's Poisson is a signed int, have to conform to that!

* Fix minor warnings in BayesIntervalCalc

* Remove unused ContainedType typedef

* singed/unsigned in SystematicManager.cpp

* signed.unsigned in BinnedEDShrinker solved w/ static_cast

* signed/usigned in DistFiller

* Remove unused param in DistTools::ToTH2D()

* Remove unused param in BinnedEDGenerator::GetPdfs()

* Fix wrong type in StatatisticSumTest

* Fix type in VaryingCDFTest.cpp

* Fix signed/unsigned in test DataSetGenerator

* fix signed/unsigned in ParameterManagerTest.cpp

* Fill out silly methods for FakeSystematic

* Fix logic in VaryingCDF::GetParameter().

* Remove unused code in ROOTMultiPlot::Construct()

* Deal with signed/unsigned warningsin DataSetGenerator via static_casts

* Remove -fdiangostics-color flag

* Revert a use of auto in BinnedNLLH.cpp

* Revert a use of auto in ChiSquare.cpp

* Reorder default constructor for AutoCorrelationCalc, MCMCSamples

* Reorder constructor for MCMC

* Reorder constructors for BoolCut and BoxCut

* Change fQlen and contructor input of AutoCorrelationCalc to be size_t

* AutoCorrelationCalc fix pt.2

* Remove redundant input param for MCSampler::CorrectAccParam() and derived classes

* Removal of signed/unsigned errors in DataSetGenerator.cpp

* fix bug

* Enforce counts in PossonFluctuatedDataSet() to be unsigned.

+++ src/fitutil/BinnedEDShrinker.cpp
@@ -111,7 +111,7 @@ BinnedEDShrinker::ShrinkDist(const BinnedED& dist_) const{
             }
 
             // bins in the upper buffer have i > number of bins in axis j. Do the same
-            if (offsetIndex >= newAxes.GetAxis(j).GetNBins()){
+            if (offsetIndex >= static_cast<int>(newAxes.GetAxis(j).GetNBins())){
                 offsetIndex = newAxes.GetAxis(j).GetNBins() - 1;
 
                 if (!fUsingOverflows)
